changed return for getDuration to unsigned long

+++ Libraries/Sensor/Sensor.hpp
@@ -2,13 +2,14 @@
 #define SENSOR_HPP
 
 class Sensor {
-	public:
-		int	getDuration();
-		Sensor(const unsigned long *_globMillis);
-		Sensor();
-		~Sensor();
-	private:
-		const unsigned long	*globMillis;
+public:
+    unsigned long getDuration();
+    Sensor(const unsigned long* _globMillis);
+    Sensor();
+    ~Sensor();
+
+private:
+    const unsigned long* globMillis;
 };
 
 #endif
